+++ CHANGELOG.md
@@ -1,11 +1,15 @@
 # Change Log
 
 ## 0.2.8
+- Updated docblocks, and added examples
 - Fixed `array.split()`
-- Added examples to docblocks
 - Removed `client.script()`, `client.size()`, & `client.stylesheet()`
 - Defaulting second parameter of `client.request()` to `GET`
 - Exposed `keigai.util.repeat()`
+- Added `addEventListener()`, `removeEventListener()`, & `removeListener()` facades on prototypes for consistent APIs
+- Created `DataGrid.add()`, `DataGrid.remove()`, & `DataGrid.update()` as a simple abstraction of the underlying `DataStore`
+- Created `DataList.add()`, `DataList.remove()`, & `DataList.update()` as a simple abstraction of the underlying `DataStore`
+- Created `DataStore.addEventListener()`, & `DataStore.removeEventListener()`
 
 ## 0.2.7
 - Fixed alignment mistakes